Rename Node to GenricNode and SyntaxNode to Node
